move Activation to User.Activation
